# Optimization for general ND copies