- Add specializations for more work per thread to amortize cost of indexing
- Use 2-array index computation for gg copy to save a bunch of divides (helps a lot)

Benchmarking general and gg copy for 5D arrays:

bench | pre | post
---- | ---- | ----
general | 118.26 ms  |  79.3 ms
gg | 230.2 ms |  107.0 ms

Benchmark below:

```python
D = 64

a = mx.zeros((D, D, D, D, D))
a = mx.transpose(a, (1, 0, 4, 2, 3))

def fn_g(a):
    return [a[..., ::-1] for _ in range(5)]

def fn_gg(a):
    return [mx.pad(a, ((0, 0), (0, 0), (0, 0), (0, 0), (4, 0))) for _ in range(5)]
```